pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "wagtail-ai"
authors = [{name = "Tom Usher", email = "tom@tomusher.com"}]
description = "Get a hand writing your content with AI super powers!"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Django",
    "Framework :: Django :: 3.0",
    "Framework :: Django :: 3.1",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Wagtail",
    "Framework :: Wagtail :: 4",
]
dynamic = ["version"]
dependencies = [
    "Django>=3.0",
    "Wagtail>=4.0",
    "langchain>=0.0.106",
    "tiktoken>=0.3.3"
]
[project.optional-dependencies]
numpy = [
    "numpy>=1.24.2",
]
qdrant = [
    "qdrant-client>=1.1.4",
]
weaviate = [
    "weaviate-client>=3.15.6",
]
testing = [
    "dj-database-url==1.2.0",
    "pre-commit>=2.21.0",
    "pytest>=7.2.2",
    "pytest-django>=4.5.2",
    "wagtail-factories>=4.0.0",
    "factory-boy>=3.2.1"
]
docs = [
    "mkdocs>=1.4.2",
    "mkdocs-material>=9.1.3",
    "mkdocs-awesome-pages-plugin>=2.8.0"
]

[project.urls]
Home = "https://github.com/tomusher/wagtail-ai"

[tool.ruff]
select = ["F", "E", "C90", "I", "B", "DJ", "RUF", "TRY", "C4"]
ignore = ["TRY003", "E501"]

[tool.flit.module]
name = "wagtail_ai"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "testapp.settings"
pythonpath = ["./", "./tests"]

[tool.flit.sdist]
include = [
    "wagtail_ai/static"
]
exclude = [
    "wagtail_ai/static_src",
    "wagtail_ai/test",
    "wagtail_ai/static/wagtail_ai/js/.gitignore",
    "tests",
    "testmanage.py",
    ".*",
    "*.js",
    "*.json",
    "*.ini",
    "*.yml"
]
